requirements.txt
saltair==4.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.1
asttokens==2.0.5
attrs==22.1.0
backcall==0.2.0
bleach==5.0.1
blinker==1.5
cachetools==5.2.0
calmap==0.0.9
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.4
commonmark==0.9.1
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
fastjsonschema==2.16.1
fonttools==4.33.3
geographiclib==1.52
geopy==2.2.0
gitdb==4.0.9
GitPython==3.1.27
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.12.0
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.9.1
jupyter-client==7.3.1
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
nltk==3.7
notebook==6.4.12
numpy==1.22.3
openpyxl==3.0.10
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.1
plotly==5.9.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.0
pure-eval==0.2.2
pyarrow==9.0.0
pycparser==2.21
pydeck==0.7.1
Pygments==2.12.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pyzmq==22.3.0
regex==2022.7.25
requests==2.28.1
rich==12.5.1
scikit-learn==1.1.1
scipy==1.8.1
seaborn==0.11.2
semver==2.13.0
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
sodapy==2.1.1
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.2.0
streamlit==1.11.1
tenacity==8.0.1
terminado==0.15.0
threadpoolctl==3.1.0
tinycss2==1.1.1
toml==0.10.2
toolz==0.12.0
tornado==6.1
tqdm==4.64.0
traitlets==5.3.0
typing_extensions==4.3.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.11
validators==0.20.0
watchdog==2.1.9
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==6.2.0
widgetsnbextension==3.6.1
xlrd==2.0.1
zipp==3.8.1
